tests/wpt/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/002.html

<!doctype html>
<title>Multiple simultaneous navigations</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="about:blank"></iframe>
<script>
var t = async_test();
onload = t.step_func(function() {
  var iframe = document.getElementsByTagName("iframe")[0];

  setTimeout(t.step_func(function() {
    iframe.src = "002-1.html?pipe=trickle(d1)";
    setTimeout(t.step_func(function(){iframe.src = "002-2.html"}), 500);
  }), 100);
});

onmessage = t.step_func(function(e) {
  assert_equals(e.data, "002-2");
  t.done();
});

</script>
